app/requirements.txt
atomicwrites==1.2.1
attrs==18.2.0
Click==7.0
Flask==1.0.2
Flask-JWT==0.3.2
Flask-JWT-Extended==3.14.0
Flask-Selfdoc==1.1.0
itsdangerous==1.1.0
Jinja2==2.10
MarkupSafe==1.1.0
more-itertools==4.3.0
mysqlclient==1.3.14
pluggy==0.8.0
py==1.7.0
pyaml==18.11.0
PyJWT==1.4.2
pytest==4.0.2
PyYAML==3.13
six==1.12.0
SQLAlchemy==1.2.15
Werkzeug==0.14.1